clean up 3.0

+++ app/src/main/java/org/mobilitychoices/activities/LoginActivity.java
@@ -91,16 +91,17 @@ public class LoginActivity extends AppCompatActivity implements LoaderCallbacks<
         new MeTask(new MeTask.IMeCallback() {
             @Override
             public void done(Response<Object> response) {
-                Toast.makeText(LoginActivity.this.getApplicationContext(), String.valueOf(response.getCode()), Toast.LENGTH_SHORT).show();
-                if (response.getCode() == 200) {
-                    Intent intent = new Intent(LoginActivity.this, MainActivity.class);
-                    LoginActivity.this.startActivity(intent);
-                    LoginActivity.this.finish();
-                } else if (response.getCode() == 401) {
-                    //remove token from preferences
-                    SharedPreferences.Editor editor = sharedPref.edit();
-                    editor.clear();
-                    editor.apply();
+                if(response != null){
+                    if (response.getCode() == 200) {
+                        Intent intent = new Intent(LoginActivity.this, MainActivity.class);
+                        LoginActivity.this.startActivity(intent);
+                        LoginActivity.this.finish();
+                    } else if (response.getCode() == 401) {
+                        //remove token from preferences
+                        SharedPreferences.Editor editor = sharedPref.edit();
+                        editor.clear();
+                        editor.apply();
+                    }
                 }
             }
         }, token).execute();
